firmware: move led logic to leds.h|c
