+++ firmware/hackrf-common.cmake
@@ -172,6 +172,7 @@ macro(DeclareTargets)
 		${PATH_HACKRF_FIRMWARE_COMMON}/cpu_clock.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/delay.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/hackrf_core.c
+		${PATH_HACKRF_FIRMWARE_COMMON}/leds.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/sgpio.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/rf_path.c
 		${PATH_HACKRF_FIRMWARE_COMMON}/si5351c.c
