+++ firmware/hackrf_usb/hackrf_usb.c
@@ -36,6 +36,7 @@
 #include <fixed_point.h>
 #include <hackrf_core.h>
 #include <hackrf_ui.h>
+#include <leds.h>
 #include <operacake.h>
 #include <platform_detect.h>
 #include <power.h>
